Split logic and main program into separate packages
Move the logic from the main package into it's own ccvm package. In
order to support this some functions were exported from the ccvm package
for the key operations on the VM.
